# docs/developer-docs/public/indexer/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "tari_indexer",
    "description": "The Tari indexer implementation",
    "contact": {
      "name": "The Tari Development Community"
    },
    "license": {
      "name": "BSD-3-Clause",
      "identifier": "BSD-3-Clause"
    },
    "version": "0.37.0"
  },
  "paths": {
...
    "/utxos": {
      "get": {
        "tags": [
          "handlers::utxos"
        ],
        "description": "List full UTXO data",
        "operationId": "list_utxos",
        "responses": {
          "200": {
            "description": "List of UTXOs",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListUtxosResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request parameters",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Failed to list UTXOs",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/utxos/fetch": {
      "post": {
        "tags": [
          "handlers::utxos"
        ],
        "description": "Gets full UTXO data for a list of UTXO IDs",
        "operationId": "fetch_utxos",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GetUtxosRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "UTXO details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetUtxosResponse"
                }
              }
            }
          },
          "400": {
            "description": "Too many UTXOs requested or invalid parameters",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Failed to fetch UTXOs",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/utxos/stream": {
      "post": {
        "tags": [
          "handlers::utxos"
        ],
        "description": "Streams UTXO updates",
        "operationId": "stream_utxo_updates",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GetUtxoUpdatesRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Stream of UTXO updates"
          },
          "400": {
            "description": "Invalid request parameters",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Failed to start UTXO update stream",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string"
          }
        }
      },
...
      "GetUtxoUpdatesRequest": {
        "type": "object",
        "required": [
          "shard_state_versions",
          "resource_address",
          "per_shard_limit"
        ],
        "properties": {
          "from_epoch": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "per_shard_limit": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "resource_address": {
            "type": "string"
          },
          "shard_state_versions": {
            "type": "array",
            "items": false,
            "prefixItems": [
              {
                "type": "integer",
                "format": "int32",
                "minimum": 0
              },
              {
                "type": "integer",
                "format": "int64",
                "minimum": 0
              }
            ]
          },
          "unspent_only": {
            "type": "boolean"
          }
        }
      },
      "GetUtxosRequest": {
        "type": "object",
        "required": [
          "tag_and_nonce_pairs",
          "resource_address"
        ],
        "properties": {
          "resource_address": {
            "type": "string"
          },
          "tag_and_nonce_pairs": {
            "type": "array",
            "items": {
              "type": "array",
              "items": false,
              "prefixItems": [
                {
                  "type": "integer",
                  "format": "int32",
                  "minimum": 0
                },
                {
                  "type": "string"
                }
              ]
            }
          }
        }
      },
      "GetUtxosResponse": {
        "type": "object",
        "required": [
          "utxos"
        ],
        "properties": {
          "utxos": {
            "type": "array",
            "items": {
              "type": "array",
              "items": false,
              "prefixItems": [
                {
                  "type": "string"
                },
                {
                  "type": "object"
                }
              ]
            }
          }
        }
      },
...
      "ListUtxosResponse": {
        "type": "object",
        "required": [
          "utxos"
        ],
        "properties": {
          "utxos": {
            "type": "array",
            "items": {
              "type": "array",
              "items": false,
              "prefixItems": [
                {
                  "type": "string"
                },
                {
                  "type": "object"
                }
              ]
            }
          }
        }
      },
...
    }
  }
}